composer: remove git hook
This package is used to validate de code style but, this validation is
done before commits and against PEAR standard. This isn't working well
in development environment and on travis-ci environments. So, this PR
intent to remove this package from project to avoid problems on
travis-ci and, enable the implementation to another tools to made
validate coding standards like codeclimante etc.